refactor: file names and folders
