+++ src/components/actions/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable mobx/missing-observer */
 import { Moon, Sun } from 'lucide-react';
 
-import { useTheme } from '@/components/theme-provider';
+import { useTheme } from '@/components/actions/ThemeProvider';
 import { Button } from '@/components/ui/button';
 import {
   DropdownMenu,
